+++ code/src/cljs/sixsq/nuvla/ui/edge_detail/spec.cljs
@@ -19,6 +19,7 @@
 (s/def ::nuvlabox-cluster (s/nilable any?))
 (s/def ::nuvlabox-not-found? boolean?)
 (s/def ::nuvlabox-playbooks (s/nilable any?))
+(s/def ::nuvlabox-emergency-playbooks (s/nilable any?))
 
 
 (s/def ::db (s/keys :req [::nuvlabox
@@ -36,7 +37,8 @@
                           ::join-token
                           ::nuvlabox-cluster
                           ::nuvlabox-not-found?
-                          ::nuvlabox-playbooks]))
+                          ::nuvlabox-playbooks
+                          ::nuvlabox-emergency-playbooks]))
 
 
 (def defaults {::nuvlabox                     nil
@@ -54,4 +56,5 @@
                ::join-token                   nil
                ::nuvlabox-cluster             nil
                ::nuvlabox-not-found?          false
-               ::nuvlabox-playbooks           nil})
+               ::nuvlabox-playbooks           nil
+               ::nuvlabox-emergency-playbooks nil})