+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
   - NuvlaBox - new options for opting in on host-level management, at NuvlaBox creation time
   - NuvlaBox - new tab for managing playbooks
+  - NuvlaBox - new buttons for playbook management operations
 
 ### Changed